+++ tests/conftest.py
@@ -4,6 +4,8 @@ import uuid
 
 import requests
 
+from aws_xray_sdk.core import xray_recorder
+
 from insanic import Insanic
 from insanic.authentication import handlers
 from insanic.choices import UserLevels
@@ -12,6 +14,7 @@ from insanic.models import User
 
 from pytest_redis import factories
 
+xray_recorder.configure(sampling=False, context_missing="LOG_ERROR")
 settings.configure(SERVICE_NAME="insanic", GATEWAY_REGISTRATION_ENABLED=False, MMT_ENV="test",
                    TRACING_ENABLED=False)